requirements.txt
streamlit
pandas
fbprophet
plotly
